+++ python-packages/securesync/devices/api_client.py
@@ -1,17 +1,12 @@
 """
 """
-import logging
 import json
-import re
-import requests
 import urllib
-import urllib2
 
-from django.conf import settings
 from django.core.urlresolvers import reverse
 
 from .models import *
-from .. import engine
+from ..engine.utils import serialize, deserialize
 from ..api_client import BaseClient
 from fle_utils.internet.functions import set_query_params
 
@@ -93,7 +88,7 @@ class RegistrationClient(BaseClient):
         # the other side will have to reconstruct the chain from the object list
         object_list = [own_device] + chain_of_trust.objects()
         r = self.post("register", {
-            "client_device": engine.serialize(object_list, ensure_ascii=False),
+            "client_device": serialize(object_list, ensure_ascii=False),
         })
 
         # Failed to register with any certificate
@@ -102,7 +97,7 @@ class RegistrationClient(BaseClient):
 
         # When we register, we should receive the model information we require.
         #   Make sure to deserialize for our version.
-        return (engine.deserialize(r.content, dest_version=own_device.get_version()), r)
+        return (deserialize(r.content, dest_version=own_device.get_version()), r)
 
 
     def register_via_preregistered_key(self):
@@ -116,7 +111,7 @@ class RegistrationClient(BaseClient):
         # Note that (currently) this should never fail--the central server (which we're sending
         #   these objects to) should always have a higher version.
         r = self.post("register", {
-            "client_device": engine.serialize([own_device], ensure_ascii=False),
+            "client_device": serialize([own_device], ensure_ascii=False),
         })
 
         # If they don't understand, our assumption is broken.
@@ -131,7 +126,7 @@ class RegistrationClient(BaseClient):
         else:
             # Save to our local store.  By NOT passing a src_version,
             #   we're saying it's OK to just store what we can.
-            return engine.deserialize(r.content, src_version=None, dest_version=own_device.get_version())
+            return deserialize(r.content, src_version=None, dest_version=own_device.get_version())
 
     def get_registration_url(self, **kwargs):
