+++ python-packages/securesync/engine/models.py
@@ -17,7 +17,7 @@ from django.dispatch import receiver
 from django.utils.text import compress_string
 from django.utils.translation import ugettext_lazy as _
 
-from . import add_syncing_models
+from .utils import add_syncing_models
 from .. import ID_MAX_LENGTH, IP_MAX_LENGTH
 from fle_utils.config.models import Settings
 from fle_utils.django_utils.debugging import validate_via_booleans
