+++ python-packages/securesync/management/commands/retrypurgatory.py
@@ -1,8 +1,6 @@
-import json
+from django.core.management.base import BaseCommand
 
-from django.core.management.base import BaseCommand, CommandError
-
-from ... import engine
+from ...engine.utils import save_serialized_models
 from ...models import ImportPurgatory
 
 
@@ -29,7 +27,7 @@ class Command(BaseCommand):
             # Serialized version is ourselves (or an earlier version of ourselves),
             #   so say so explicitly to make sure errors get bubbled up to us properly.
             try:
-                unsaved = engine.save_serialized_models(data=purgatory)["unsaved_model_count"]
+                unsaved = save_serialized_models(data=purgatory)["unsaved_model_count"]
                 if not unsaved:
                     self.stdout_writeln("\t%s :)" % (("All models were saved successfully!")))
                 else:
